refactor: style, comments, and file structure
