+++ src/hooks/components/index.js
@@ -1,0 +1,1 @@
+export { default as Initializer } from './initializer'
