+++ src/index.js
@@ -1,6 +1,6 @@
-import drizzleConnect from "./drizzleConnect.js";
-import DrizzleProvider from "./DrizzleProvider.js";
-import DrizzleContext from "./DrizzleContext";
-import * as drizzleReactHooks from "./drizzleReactHooks"
+import drizzleConnect from './drizzleConnect.js'
+import DrizzleProvider from './DrizzleProvider.js'
+import DrizzleContext from './DrizzleContext'
+import * as drizzleReactHooks from './hooks'
 
-export { drizzleConnect, DrizzleProvider, DrizzleContext, drizzleReactHooks };
+export { drizzleConnect, DrizzleProvider, DrizzleContext, drizzleReactHooks }
